move chart code
